feat(jit): separate diagnostics assembly from runtime
